API Reference

MarketData

FinancialDSL.MarketData.AbstractMarketDataProvider — Type

Interface

MarketData.get_value(provider, serie_id, at, asof::Date; locf::Bool=false) :: Union{Missing, T}

MarketData.get_serie_currency(provider, serie_id) :: Union{Nothing, Currecies.Currency}

MarketData.has_serie(provider, serie_id) :: Bool

Arguments

at : a key to the value of the time series.

asof : the closing date at which the time series is being assessed. This is usually set to the pricing date.

locf : last observation carried forward. Use locf=true to repeat the latest observation prior to at.

Provided methods

MarketData.has_value(provider, serie_id, at, asof::Date; locf::Bool=false) :: Bool

MarketData.get_cash(provider, serie_id, at, asof::Date; locf::Bool=false) :: Union{Missing, Currencies.Cash}

MarketData.assert_has_serie(provider, serie_id)

FinancialDSL.MarketData.get_value — Function

get_value(provider, serie_id, at, asof::Date; locf::Bool=false) :: Union{Missing, T}

Returns a value for serie_id at state (or date) at.

Arguments

at : a key to the value of the time series.

asof : the date at which the time series is being assessed.

locf : last observation carried forward. Use locf=true to repeat the latest observation prior to at.

This method should error if provider does not know about serie_id. Use MarketData.assert_has_serie for that.

FinancialDSL.MarketData.get_serie_currency — Function

get_serie_currency(provider, serie_id) :: Union{Nothing, Currecies.Currency}

Returns the currency in which serie_id is expressed. Returns nothing if serie_id is not based on a currency.

This method should error if provider does not know about serie_id. Use MarketData.assert_has_serie for that.

FinancialDSL.MarketData.EmptyMarketDataProvider — Type

A Market Data provided that has no data. Useful when calculation does not depend on Market Data.

Core

FinancialDSL.Core.Observable — Type

Represents a time-varying quantity of type {T}. So, its value is indexed by a date or datetime.

FinancialDSL.Core.PricingDate — Type

PricingDate <: Observable{Date}

Represents the pricing date

FinancialDSL.Core.Konst — Type

The value of Konst(x) is always x at any date.